Server/requirements.txt
"kivy[base]"==2.2.1
kivy_examples==2.2.1
requests